+++ tests/test_yaralyze.py
@@ -1,6 +1,7 @@
 """
 Tests for invoking yaralyze script from shell.
 """
+from functools import partial
 from math import isclose
 from os import environ, path, remove
 from subprocess import CalledProcessError, check_output
@@ -38,25 +39,21 @@ def test_too_many_rule_args(il_tulipano_path, tulips_yara_path):
         _run_with_args(il_tulipano_path, '-Y', tulips_yara_path, '-hex', HEX_STRING)
 
 
-def test_yaralyze(il_tulipano_path, tulips_yara_path):
-    # yaralyze -y tests/file_fixtures/tulips.yara tests/file_fixtures/il_tulipano_nero.txt
-    with_yara_file_output = _run_with_args(il_tulipano_path, '-Y', tulips_yara_path)
-    # yaralyze -dir tests/file_fixtures/ tests/file_fixtures/il_tulipano_nero.txt
-    with_dir_output = _run_with_args(il_tulipano_path, '-dir', path.dirname(tulips_yara_path))
-    counts = [line_count(output) for output in [with_yara_file_output, with_dir_output]]
-
-    for c in counts:
-        assert isclose(c, EXPECTED_LINES, rel_tol=CLOSENESS_THRESHOLD), f"{c} is too far from {EXPECTED_LINES}"
+def test_yaralyze_with_files(il_tulipano_path, tulips_yara_path):
+    """
+    Check output of:
+        yaralyze -Y tests/file_fixtures/tulips.yara tests/file_fixtures/il_tulipano_nero.txt
+        yaralyze -dir tests/file_fixtures/ tests/file_fixtures/il_tulipano_nero.txt
+    """
+    test_line_count = partial(_assert_output_line_count_is_close, EXPECTED_LINES, il_tulipano_path)
+    test_line_count('-Y', tulips_yara_path)
+    test_line_count('-dir', path.dirname(tulips_yara_path))
 
 
 def test_yaralyze_with_patterns(il_tulipano_path, binary_file_path, tulips_yara_regex):
-    # yaralyze -r 'tulip.{1,2500}tulip' tests/file_fixtures/il_tulipano_nero.txt
-    with_pattern_output = _run_with_args(il_tulipano_path, '-re', tulips_yara_regex)
-    assert line_count(with_pattern_output) == 942
-    with_pattern_output = _run_with_args(binary_file_path, '-re', '3Hl0')
-    assert line_count(with_pattern_output) == 83
-    with_pattern_output = _run_with_args(binary_file_path, '-hex', HEX_STRING)
-    assert line_count(with_pattern_output) == 90
+    _assert_output_line_count_is_close(942, il_tulipano_path, '-re', tulips_yara_regex)
+    _assert_output_line_count_is_close(83, binary_file_path, '-re', '3Hl0')
+    _assert_output_line_count_is_close(90, binary_file_path, '-hex', HEX_STRING)
 
 
 def test_file_export(binary_file_path, tulips_yara_path, tmp_dir):
@@ -64,14 +61,21 @@ def test_file_export(binary_file_path, tulips_yara_path, tmp_dir):
     rendered_files = files_in_dir(tmp_dir)
     assert len(rendered_files) == 3
     file_sizes = [path.getsize(f) for f in rendered_files]
-    assert_array_is_close(sorted(file_sizes), [40867, 69127, 216719])
+    _assert_array_is_close(sorted(file_sizes), [40867, 69127, 216719])
 
     for file in rendered_files:
         remove(file)
 
 
-def assert_output_line_count(shell_cmd: list, expected_line_count: int):
-    _assert_line_count_within_range(expected_line_count, check_output(shell_cmd).decode())
+def _assert_array_is_close(_list1, _list2):
+    for i, item in enumerate(_list1):
+        if not isclose(item, _list2[i], rel_tol=CLOSENESS_THRESHOLD):
+            assert False, f"File size of {item} too far from {_list2[i]}"
+
+
+def _assert_output_line_count_is_close(expected_line_count: int, file_to_scan: str, *args) -> None:
+    output_line_count = line_count(_run_with_args(file_to_scan, *args))
+    assert isclose(expected_line_count, output_line_count, rel_tol=CLOSENESS_THRESHOLD)
 
 
 def _run_with_args(file_to_scan, *args) -> str:
@@ -79,14 +83,9 @@ def _run_with_args(file_to_scan, *args) -> str:
     return check_output([YARALYZE, file_to_scan, *args], env=environ).decode()
 
 
-def _assert_line_count_within_range(line_count, text):
-    lines_in_text = len(text.split("\n"))
+def _assert_line_count_within_range(expected_line_count, text):
+    lines_in_text = line_count(text)
 
-    if not isclose(line_count, lines_in_text, rel_tol=CLOSENESS_THRESHOLD):
+    if not isclose(expected_line_count, lines_in_text, rel_tol=CLOSENESS_THRESHOLD):
         console.print(text)
         raise AssertionError(f"Expected {line_count} +/- but found {lines_in_text}")
-
-def assert_array_is_close(_list1, _list2):
-    for i, item in enumerate(_list1):
-        if not isclose(item, _list2[i], rel_tol=CLOSENESS_THRESHOLD):
-            assert False, f"File size of {item} too far from {_list2[i]}"